backend/requirements.txt
flask
paho-mqtt
werkzeug
python-dotenv
flask-sqlalchemy
flask-cors
qrcode
